Fix quality printing

+++ src/outpainting.py
@@ -73,7 +73,7 @@ class Outpainting:
 
             # check quality
             quality = self.aesthetic_predictor.eval_image(generated_image)
-            print("Quality: " + str(int(quality)))
+            print("Quality: " + str(float(quality)))
             quality_threshold -= quality_step
 
         # generate transition
